pyproject.toml
[project]
name = "torken-usage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai",
    "python-dotenv"
]
